Add Logoutbutton to main dashboard toolbar
